Split code into smaller files
