+++ src/mf2.ts
@@ -3,14 +3,10 @@ import { mf2 } from 'microformats-parser';
 
 import fetch from 'node-fetch';
 import { URL } from 'url';
-import { isUri } from 'valid-url';
 
 import {
-  AuthorInfo,
-  ObjectWithStringValue,
   ParsedDocumentFetchFn,
   PartialPost,
-  PostType,
   SimplifiedCite,
   SimplifiedEntry,
   SimplifiedEvent,
@@ -20,17 +16,8 @@ import {
 
 import urljoin = require('url-join');
 import { MicroformatProperty, MicroformatRoot, Html, ParsedDocument, MicroformatProperties } from './types/microformat-parser';
-
-function is_microformat_root(p: MicroformatProperty | string): p is MicroformatRoot {
-  return (p as MicroformatRoot).properties !== undefined;
-}
-
-function is_obj_with_string_value(p: unknown): p is ObjectWithStringValue {
-  return (
-    (p as ObjectWithStringValue).value !== undefined &&
-    typeof (p as ObjectWithStringValue).value === 'string'
-  );
-}
+import { find_author } from './author';
+import { find_first_entry, get_plain_text, is_microformat_root, is_name_a_title } from './utils';
 
 function is_html(p: MicroformatProperty): p is Html {
   return p !== undefined && (p as Html).html !== undefined;
@@ -51,427 +38,6 @@ export async function parse_mf2(url: string): Promise<ParsedDocument> {
   const page: string = await response.text();
   const baseUrl = get_base_url(url);
   return mf2(page, { baseUrl });
-}
-
-/**
- * Find the first interesting h-* object in BFS-order
- *
- * @param parsed an mf2 parsed object
- * @param types array of types to look for
- * @return first h-* types that matches one of the array types
- *
- */
-export function find_first_entry(parsed: ParsedDocument, types: PostType[]): MicroformatRoot {
-  return find_all_entries_gen(parsed, types, false).next().value;
-}
-
-/**
- * Find all h-* objects of a given type in BFS-order. Traverses the
- * top-level items and their children and descendents. Includes property
- * values (e.g. finding all h-cards would not find values of
- * "p-author h-card") only if `include_properties` is True.
-
- * @param parsed an mf2 parsed object
- * @param types array of types to look for
- * @param include_properties if true we look at properties as well
- * @returns all matching objects
- */
-export function find_all_entries(
-  parsed: ParsedDocument,
-  types: PostType[],
-  include_properties = false
-): MicroformatRoot[] {
-  return [...find_all_entries_gen(parsed, types, include_properties)];
-}
-
-function* find_all_entries_gen(
-  parsed: ParsedDocument,
-  types: PostType[],
-  include_properties: boolean
-): Generator<MicroformatRoot> {
-  const queue: MicroformatRoot[] = [...parsed.items];
-  while (queue.length > 0) {
-    const item = queue.shift();
-    if (!item) {
-      throw new Error('item is undefined');
-    }
-    const item_types = item.type || [];
-    if (types.some((h_class) => _.includes(item_types, h_class))) {
-      yield item;
-    }
-    queue.push(...(item.children || []));
-    if (include_properties) {
-      const rootList = _.flatten(Object.values(item.properties || {})).filter(is_microformat_root);
-      queue.push(...rootList);
-    }
-  }
-}
-
-/**
- * Get the first value in a list of values that we expect to be plain-text.
- * If it is an object, then return the value of "value".
- *
- * @param values a list of values
- * @param strip true if we should strip the plaintext value
- * @return the text value or null
- */
-function get_plain_text(values: MicroformatProperty[], strip = false): string | null {
-  if (_.size(values) === 0) {
-    return null;
-  }
-
-  const value = values[0];
-  let v: string | null;
-  if (is_obj_with_string_value(value)) {
-    v = value.value;
-  } else if (typeof value === 'string') {
-    v = value;
-  } else {
-    v = null;
-  }
-  return v && strip ? v.trim() : v;
-}
-
-/**
- * Parse the value of a u-author property, can either be a compound
- * h-card or a single name or url.
- * 
- * @param obj the mf2 property value, either an object or a string
-   @return an object containing the author's name, photo, and url
- */
-function parse_author(obj: string | MicroformatRoot): AuthorInfo {
-  const result: AuthorInfo = {};
-  if (is_microformat_root(obj)) {
-    const names = obj.properties.name as string[];
-    if (_.size(names) > 0) {
-      result.name = names[0];
-    }
-
-    const photos = obj.properties.photo;
-    if (_.size(photos) > 0) {
-      const photo = get_plain_text(photos);
-      if (photo) {
-        result.photo = photo;
-      }
-    }
-
-    const urls = obj.properties.url as string[];
-    if (_.size(urls) > 0) {
-      result.url = urls[0];
-    }
-  } else if (obj) {
-    if (obj.startsWith('http://') || obj.startsWith('https://')) {
-      result.url = obj;
-    } else {
-      result.name = obj;
-    }
-  }
-  return result;
-}
-
-function urlEqual(url1: string, url2: string): boolean {
-  const _url1 = url1.endsWith('/') ? url1.slice(0, -1) : url1;
-  const _url2 = url2.endsWith('/') ? url2.slice(0, -1) : url2;
-  return _url1 === _url2;
-}
-
-/**
- * Use the authorship discovery algorithm
- * https://indiewebcamp.com/authorship to determine an h-entry's
- * author.
- *
- * @param parsed an mf2 parsed object
- * @param hentry optional, the h-netry we're examining, if omitted we'll just use the first one
- * @param fetch_author if true we will follow author page URLs
- * @return a promise for an object containing author's name, photo, url
- */
-export async function find_author(
-  parsed: ParsedDocument,
-  hentry: MicroformatRoot | null,
-  fetch_mf2_func: ParsedDocumentFetchFn | null
-): Promise<AuthorInfo | null> {
-  function find_hentry_author(hentry: MicroformatRoot) {
-    const vals = hentry.properties.author || [];
-    if (_.size(vals) === 0) {
-      return null;
-    }
-    return parse_author(vals[0] as MicroformatRoot);
-  }
-
-  function find_parent_hfeed_author(hentry: MicroformatRoot) {
-    const hfeeds = find_all_entries_gen(parsed, ['h-feed'], false);
-    for (const hfeed of hfeeds) {
-      const feed_children = hfeed.children || [];
-      if (feed_children.includes(hentry)) {
-        // not the hentry, but this works
-        return find_hentry_author(hfeed);
-      }
-    }
-  }
-
-  hentry = hentry || find_first_entry(parsed, ['h-entry']);
-  if (!hentry) {
-    return null;
-  }
-
-  // 3. if the h-entry has an author property, use that
-  // 4. otherwise if the h-entry has a parent h-feed with author property,
-  //    use that
-  const author = find_hentry_author(hentry) || find_parent_hfeed_author(hentry);
-  let author_page;
-  if (author) {
-    // 5.2 otherwise if author property is an http(s) URL, let the
-    //     author-page have that URL
-    if (_.isEqual(Object.keys(author), ['url'])) {
-      author_page = author['url'];
-    }
-    // 5.1 if it has an h-card, use it, exit.
-    // 5.3 otherwise use the author property as the author name,
-    //     exit.
-    else {
-      return author;
-    }
-  }
-
-  // 6. if there is no author-page and the h-entry's page is a permalink page
-  if (!author_page) {
-    // 6.1 if the page has a rel-author link, let the author-page's
-    //     URL be the href of the rel-author link
-    const rel_authors = (parsed.rels || {}).author || [];
-    if (_.size(rel_authors) > 0) {
-      author_page = rel_authors[0];
-    }
-  }
-  if (author_page) {
-    if (!fetch_mf2_func) {
-      return {
-        url: author_page,
-      };
-    }
-
-    // 7.1 get the author-page from that URL and parse it for microformats2
-    parsed = await fetch_mf2_func(author_page);
-    const hcards = find_all_entries(parsed, ['h-card'], true);
-
-    // 7.2 if author-page has 1+ h-card with url == uid ==
-    //     author-page's URL, then use first such h-card, exit.
-    for (const hcard of hcards) {
-      const hcard_url = get_plain_text(hcard.properties.url);
-      const hcard_uid = get_plain_text(hcard.properties.uid);
-      if (hcard_url && hcard_uid && hcard_url === hcard_uid && urlEqual(hcard_url, author_page)) {
-        return parse_author(hcard);
-      }
-    }
-
-    // 7.3 else if author-page has 1+ h-card with url property
-    //     which matches the href of a rel-me link on the author-page
-    //     (perhaps the same hyperlink element as the u-url, though not
-    //     required to be), use first such h-card, exit.
-    const rel_mes = (parsed.rels || {}).me || [];
-    for (const hcard of hcards) {
-      const hcard_url = get_plain_text(hcard.properties.url);
-      if (hcard_url && rel_mes.includes(hcard_url)) {
-        return parse_author(hcard);
-      }
-    }
-
-    // 7.4 if the h-entry's page has 1+ h-card with url ==
-    //     author-page URL, use first such h-card, exit.
-    for (const hcard of hcards) {
-      const hcard_url = get_plain_text(hcard.properties.url);
-      if (hcard_url && urlEqual(hcard_url, author_page)) {
-        return parse_author(hcard);
-      }
-    }
-
-    // 8. otherwise no deterministic author can be found.
-    return null;
-  }
-
-  return null;
-}
-
-/**
- * Find the representative h-card for a URL
- * http://microformats.org/wiki/representative-h-card-parsing
-
- * @param parsed
- * @param source_url
- * @return the representative h-card if one is found
- */
-export function representative_hcard(
-  parsed: ParsedDocument,
-  source_url: string
-): MicroformatRoot | null {
-  const hcards = find_all_entries(parsed, ['h-card'], true);
-
-  // uid and url both match source_url
-  for (const hcard of hcards) {
-    if (
-      _.includes(hcard.properties.uid, source_url) &&
-      _.includes(hcard.properties.url, source_url)
-    ) {
-      return hcard;
-    }
-  }
-
-  // url that is also a rel=me
-  const rel_mes = (parsed.rels || {}).me || [];
-  for (const hcard of hcards) {
-    if (_.some(hcard.properties.url, (url) => _.includes(rel_mes, url))) {
-      return hcard;
-    }
-  }
-
-  // single hcard with matching url
-  let found = null;
-  let count = 0;
-  for (const hcard of hcards) {
-    if (_.includes(hcard.properties.url, source_url)) {
-      found = hcard;
-      count += 1;
-    }
-  }
-
-  return count === 1 ? found : null;
-}
-
-/**
- * Determine whether the name property represents an explicit title.
- * Typically when parsing an h-entry, we check whether p-name ==
- *  e-content (value). If they are non-equal, then p-name likely
- * represents a title.
- *
- * However, occasionally we come across an h-entry that does not
- * provide an explicit p-name. In this case, the name is
- * automatically generated by converting the entire h-entry content
- * to plain text. This definitely does not represent a title, and
- * looks very bad when displayed as such.
- *
- * To handle this case, we broaden the equality check to see if
- * content is a subset of name. We also strip out non-alphanumeric
- * characters just to make the check a little more forgiving.
- *
- * @param name the p-name property that may represent a title
- * @param content the plain-text version of an e-content property
- * @return true if the name likely represents a separate, explicit title
- *
- */
-function is_name_a_title(
-  name: string | null | undefined,
-  content: string | null | undefined
-): boolean {
-  function normalize(s: string) {
-    s = s.normalize('NFKD');
-    s = s.toLowerCase();
-    s = s.replace(
-      /(~|`|!|@|#|$|%|^|&|\*|\(|\)|{|}|\[|\]|;|:|"|'|<|,|\.|>|\?|\/|\\|\||-|_|\+|=)/g,
-      ''
-    );
-    s = s.replace(/\s+/g, '');
-    return s;
-  }
-  if (!content) {
-    return true;
-  }
-  if (!name) {
-    return false;
-  }
-  return !normalize(name).includes(normalize(content));
-}
-
-function is_prop_uri(props: MicroformatProperties, name: string): string | undefined {
-  const value = get_plain_text(props[name]);
-  return value ? isUri(value) : undefined;
-}
-
-function is_rsvp(item: MicroformatRoot): boolean {
-  const rsvp = item.properties.rsvp;
-  return (
-    rsvp &&
-    (_.includes(rsvp, 'yes') ||
-      _.includes(rsvp, 'no') ||
-      _.includes(rsvp, 'maybe') ||
-      _.includes(rsvp, 'interested'))
-  );
-}
-
-/**
- * Implementation of the post-type discovery algorithm
- * defined here https://indiewebcamp.com/post-type-discovery#Algorithm
- *
- * @param hentry mf2 item representing the entry to test
- * @return one of: 'event', 'rsvp',
- * 'reply', 'repost', 'like', 'photo','article', 'note', 'follow'
- *
- * TODO add invite, follow-of
- */
-export function post_type_discovery(item: MicroformatRoot): string {
-  if (_.includes(item.type, 'h-event')) {
-    return 'event';
-  }
-
-  if (is_rsvp(item)) {
-    return 'rsvp';
-  }
-
-  const props = item.properties;
-  const propNames = Object.keys(props);
-
-  const impliedTypes = [
-    ['repost-of', 'repost'],
-    ['like-of', 'like'],
-    ['bookmark-of', 'bookmark'],
-    ['in-reply-to', 'reply'],
-    ['video', 'video'],
-    ['photo', 'photo'],
-  ];
-
-  for (const it of impliedTypes) {
-    if (_.includes(propNames, it[0]) && is_prop_uri(props, it[0])) {
-      return it[1];
-    }
-  }
-
-  const name = get_plain_text(props.name);
-  const content = get_plain_text(props.content) || get_plain_text(props.summary);
-
-  if (content && name && is_name_a_title(name, content)) {
-    return 'article';
-  }
-
-  return 'note';
-}
-
-/**
- * Implementation of the response-type discovery algorithm
- * defined here https://www.w3.org/TR/post-type-discovery/
- *
- * @param mf2 item representing the entry to test
- * @return one of: 'rsvp','reply', 'repost', 'like', 'mention'
- */
-export function response_type_discovery(item: MicroformatRoot): string {
-  if (is_rsvp(item)) {
-    return 'rsvp';
-  }
-
-  const props = item.properties;
-  const propNames = Object.keys(props);
-
-  const impliedTypes = [
-    ['repost-of', 'repost'],
-    ['like-of', 'like'],
-    ['bookmark-of', 'bookmark'],
-    ['in-reply-to', 'reply'],
-  ];
-
-  for (const it of impliedTypes) {
-    if (_.includes(propNames, it[0]) && is_prop_uri(props, it[0])) {
-      return it[1];
-    }
-  }
-
-  return 'mention';
 }
 
 export function normalize_dt(s: string): string | null {
