+++ src/utils.ts
@@ -1,0 +1,139 @@
+import _ = require("lodash");
+import { ObjectWithStringValue, PostType } from "./mf2-models";
+import { MicroformatProperty, MicroformatRoot, ParsedDocument } from "./types/microformat-parser";
+
+function is_obj_with_string_value(p: unknown): p is ObjectWithStringValue {
+  return (
+    (p as ObjectWithStringValue).value !== undefined &&
+    typeof (p as ObjectWithStringValue).value === 'string'
+  );
+}
+
+export function is_microformat_root(p: MicroformatProperty | string): p is MicroformatRoot {
+  return (p as MicroformatRoot).properties !== undefined;
+}
+
+
+/**
+ * Get the first value in a list of values that we expect to be plain-text.
+ * If it is an object, then return the value of "value".
+ *
+ * @param values a list of values
+ * @param strip true if we should strip the plaintext value
+ * @return the text value or null
+ */
+export function get_plain_text(values: MicroformatProperty[], strip = false): string | null {
+  if (_.size(values) === 0) {
+    return null;
+  }
+
+  const value = values[0];
+  let v: string | null;
+  if (is_obj_with_string_value(value)) {
+    v = value.value;
+  } else if (typeof value === 'string') {
+    v = value;
+  } else {
+    v = null;
+  }
+  return v && strip ? v.trim() : v;
+}
+
+
+/**
+ * Determine whether the name property represents an explicit title.
+ * Typically when parsing an h-entry, we check whether p-name ==
+ *  e-content (value). If they are non-equal, then p-name likely
+ * represents a title.
+ *
+ * However, occasionally we come across an h-entry that does not
+ * provide an explicit p-name. In this case, the name is
+ * automatically generated by converting the entire h-entry content
+ * to plain text. This definitely does not represent a title, and
+ * looks very bad when displayed as such.
+ *
+ * To handle this case, we broaden the equality check to see if
+ * content is a subset of name. We also strip out non-alphanumeric
+ * characters just to make the check a little more forgiving.
+ *
+ * @param name the p-name property that may represent a title
+ * @param content the plain-text version of an e-content property
+ * @return true if the name likely represents a separate, explicit title
+ *
+ */
+export function is_name_a_title(
+  name: string | null | undefined,
+  content: string | null | undefined
+): boolean {
+  function normalize(s: string) {
+    s = s.normalize('NFKD');
+    s = s.toLowerCase();
+    s = s.replace(
+      /(~|`|!|@|#|$|%|^|&|\*|\(|\)|{|}|\[|\]|;|:|"|'|<|,|\.|>|\?|\/|\\|\||-|_|\+|=)/g,
+      ''
+    );
+    s = s.replace(/\s+/g, '');
+    return s;
+  }
+  if (!content) {
+    return true;
+  }
+  if (!name) {
+    return false;
+  }
+  return !normalize(name).includes(normalize(content));
+}
+
+/**
+ * Find the first interesting h-* object in BFS-order
+ *
+ * @param parsed an mf2 parsed object
+ * @param types array of types to look for
+ * @return first h-* types that matches one of the array types
+ *
+ */
+export function find_first_entry(parsed: ParsedDocument, types: PostType[]): MicroformatRoot {
+  return find_all_entries_gen(parsed, types, false).next().value;
+}
+
+/**
+ * Find all h-* objects of a given type in BFS-order. Traverses the
+ * top-level items and their children and descendents. Includes property
+ * values (e.g. finding all h-cards would not find values of
+ * "p-author h-card") only if `include_properties` is True.
+
+ * @param parsed an mf2 parsed object
+ * @param types array of types to look for
+ * @param include_properties if true we look at properties as well
+ * @returns all matching objects
+ */
+export function find_all_entries(
+  parsed: ParsedDocument,
+  types: PostType[],
+  include_properties = false
+): MicroformatRoot[] {
+  return [...find_all_entries_gen(parsed, types, include_properties)];
+}
+
+export function* find_all_entries_gen(
+  parsed: ParsedDocument,
+  types: PostType[],
+  include_properties: boolean
+): Generator<MicroformatRoot> {
+  const queue: MicroformatRoot[] = [...parsed.items];
+  while (queue.length > 0) {
+    const item = queue.shift();
+    if (!item) {
+      throw new Error('item is undefined');
+    }
+    const item_types = item.type || [];
+    if (types.some((h_class) => _.includes(item_types, h_class))) {
+      yield item;
+    }
+    queue.push(...(item.children || []));
+    if (include_properties) {
+      const rootList = _.flatten(Object.values(item.properties || {})).filter(is_microformat_root);
+      queue.push(...rootList);
+    }
+  }
+}
